\name{uniah-package}
\alias{uniah-package}
\docType{package}
\title{Fit Unimodal Additive Hazards Model}
\description{Nonparametric estimation of a unimodal or U-shape covariate effect for additive hazard model.}
\details{
\tabular{ll}{
Package: \tab uniah\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2022-05-27\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [aut,cre]
Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yunro Chung, Anastasia Ivanova, Jason P. Fine, Shape restricted addtive hazards model (in preparation).}
\concept{Unimodal regression}
\concept{Survival analysis}
\concept{Isotonic regression}
