\name{uniah-package}
\alias{uniah-package}
\docType{package}
\title{Fit Unimodal Additive Hazards Model}
\description{Nonparametric estimation of a unimodal or U-shape covariate effect for additive hazard model.}
\details{
\tabular{ll}{
Package: \tab uniah\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2015-12-15\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [cre], Anastasia Ivanova, Michael G. Hudgens and Jason P. Fine

Maintainer: Yunro Chung
}
\references{Yunro Chung, Anastasia Ivanova, Michael M. Hudgens, Jason P. Fine, Shape restricted addtive hazards model (in preparation).}
\keyword{Unimodal regression, Survival analysis, Constrained estimation}


